rename vsc.filesystem.quota.inodelog to vsc.filesystem.quota.inodes
